requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
requests==2.32.4
python-dotenv==1.1.0
PyJWT==2.10.1
PyMySQL==1.1.1
email-validator==2.2.0
Werkzeug==3.1.2
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
alembic==1.16.1
SQLAlchemy==2.0.35
greenlet==3.2.2
cffi==1.17.1
cryptography==45.0.4
pycparser==2.22 